[aat] Unprefix not essentially layout tables
